Checkout: Thank you page: Splitting the thank you page into separate components to make it easier to work on as a group.
This is a precursor to #1184 and #2482, so that we can work on these tasks in different PRs without conflicts.
