+++ assets/stylesheets/_components.scss
@@ -261,6 +261,7 @@
 @import 'my-sites/upgrades/checkout/style';
 @import 'my-sites/upgrades/checkout/stored-card';
 @import 'my-sites/upgrades/checkout/subscription-text';
+@import 'my-sites/upgrades/checkout-thank-you/style';
 @import 'my-sites/upgrades/domain-management/transfer/style';
 @import 'my-sites/stats/stats-list/style';
 @import 'notices/style';
