+++ assets/stylesheets/sections/_checkout.scss
@@ -771,185 +771,6 @@
 	}
 }
 
-.checkout__thank-you {
-	box-sizing: border-box;
-	position: static;
-	text-align: center;
-	width: 100%;
-
-	&.main {
-		max-width: 960px;
-	}
-
-	.thank-you-message {
-		padding-bottom: 20px;
-
-		.receipt-icon {
-			display: inline-block;
-			position: relative;
-
-			&::before {
-				@include noticon( '\f425', 70px );
-				color: lighten( $gray, 20% );
-			}
-
-			&::after {
-				@include noticon( '\f418', 18px );
-				background: #53b66c;
-				border-radius: 18px;
-				color: $white;
-				position: absolute;
-					top: 6px;
-					right: 15px;
-				width: 18px;
-				height: 18px;
-			}
-		}
-
-		> h1 {
-			color: $gray-dark;
-			font-weight: 100;
-			margin: 0;
-		}
-
-		> h2 {
-			color: $gray-dark;
-			font-weight: 200;
-			opacity: 0.6;
-		}
-	}
-
-	.try-out-message {
-		padding-top: 20px;
-
-		> h3 {
-			font-weight: 600;
-		}
-	}
-
-	.purchase-details-list {
-		list-style: none;
-	}
-
-	.get-support {
-		border-top: 1px solid lighten( $gray, 20% );
-		color: darken($gray, 10%);
-		font-size: 14px;
-
-		> h3 {
-			font-weight: 600;
-		}
-
-		> p {
-			margin-bottom: 0;
-		}
-
-		a {
-			border-bottom: 1px solid $blue-medium;
-		}
-	}
-
-	&.is-placeholder {
-		.checkout__thank-you-header,
-		.checkout__thank-you-subheader {
-			@include placeholder( 23% );
-
-			display: block;
-			margin: 0 auto;
-		}
-
-		.checkout__thank-you-header {
-			width: 50%;
-		}
-
-		.checkout__thank-you-subheader {
-			margin-top: 15px;
-			width: 40%;
-		}
-	}
-}
-
-// Mobile styles
-@include breakpoint( "<660px" ) {
-	.checkout__thank-you {
-		background: transparent;
-		padding: 14px;
-
-		.card {
-			background: none;
-			box-shadow: none;
-			padding: 0;
-		}
-
-		.thank-you-message {
-			> h1,
-			> h2 {
-				font-size: 17px;
-			}
-		}
-
-		.purchase-details-list {
-			display: inline-block;
-			margin: 0 0 10px 0;
-		}
-
-		.get-support {
-			border-bottom: 1px solid lighten( $gray, 20% );
-			padding: 20px 0;
-		}
-	}
-}
-
-// Desktop styles
-@include breakpoint( ">660px" ) {
-	.checkout__thank-you {
-		.thank-you-message {
-			border-bottom: 1px solid lighten( $gray, 30% );
-
-			> h1 {
-				font-size: 25px;
-			}
-			> h2 {
-				font-size: 20px;
-				max-width: 50%;
-				margin: auto;
-			}
-		}
-
-		.purchase-details-list {
-			margin: 0;
-			padding: 30px 0;
-		}
-
-		.purchase-detail {
-			vertical-align: top;
-		}
-
-		.purchase-detail-text {
-			&::before {
-				font-size: 70px;
-			}
-
-			> p {
-				margin: 0 50px 10px 50px;
-			}
-		}
-
-		.get-support {
-			padding: 20px;
-
-			> h3 {
-				display: inline;
-				margin-right: 15px;
-			}
-
-			> p {
-				display: inline;
-			}
-		}
-	}
-}
-
 // If there's no sidebar, we don't show the cart on the checkout page.
 @include breakpoint( "<660px" ) {
 	.secondary-cart {
