+++ client/my-sites/upgrades/checkout-thank-you/generic-details.jsx
@@ -1,0 +1,26 @@
+/**
+ * External dependencies
+ */
+import React from 'react';
+
+/**
+ * Internal dependencies
+ */
+import Button from 'components/button';
+import i18n from 'lib/mixins/i18n';
+
+const GenericDetails = ( { selectedSite } ) => {
+	return (
+		<ul className="purchase-details-list">
+			<Button href={ selectedSite.URL } primary>
+				{ i18n.translate( 'Back to my site' ) }
+			</Button>
+		</ul>
+	);
+};
+
+GenericDetails.propTypes = {
+	selectedSite: React.PropTypes.object.isRequired
+};
+
+export default GenericDetails;
