+++ client/my-sites/upgrades/checkout-thank-you/purchase-detail.jsx
