+++ client/my-sites/upgrades/controller.jsx
@@ -210,7 +210,7 @@ module.exports = {
 	},
 
 	checkoutThankYou: function( context ) {
-		var CheckoutThankYouComponent = require( './checkout/thank-you' ),
+		var CheckoutThankYouComponent = require( './checkout-thank-you' ),
 			basePath = route.sectionify( context.path ),
 			receiptId = Number( context.params.receiptId );
